refactor(engine): separate the context engine from the MCP server
Problem
    server.py was 6,748 lines holding two unrelated things: EntrolyEngine
    (2,059 lines) and create_mcp_server (3,509 lines, containing 76 nested tool
    handlers). Both defined a function named `optimize_context` -- a 419-line
    engine method and a 540-line MCP handler. That co-location is why the
    no-match contract lived on the handler while every non-MCP caller (CLI, SDK,
    proxy) ran without it: two functions with one name, and an import graph that
    could see neither, because the handlers are nested and have no importable
    identity.

Change
    entroly/engine.py now holds EntrolyEngine, its eight exclusive helpers, the
    three helpers shared with other top-level functions (moving them with it is
    what avoids a new cycle), the no-match contract and its five predicates,
    `_recall_path_prior`, and the module-level `logger`, `_EVIDENCE_WORD` and
    native-import block they depend on.

    server.py re-exports every moved name, so existing imports keep working
    unchanged -- including the tests that import `_evidence_backed`,
    `_honest_tokens_saved` and `apply_no_match_contract` from `entroly.server`.

        server.py  6,748 -> 4,238
        engine.py          2,942

Invariant
    This is a relocation, not a redesign. No behaviour changed, and
    `entroly.server.EntrolyEngine is entroly.engine.EntrolyEngine` -- one class
    re-exported, not a copy. The dependency surface was computed rather than
    guessed: 35 module-level symbols, classified into stdlib, imported classes,
    and server-local helpers, then split by whether each helper is used by the
    engine only or shared.

What this does NOT fix
    The composition problem it exposes. EntrolyEngine still leaves
    set_fast_path_router, set_journal_callback and set_crystallization_callback
    for the caller to wire, and only create_mcp_server does it, so a CLI or SDK
    engine is still assembled differently from an MCP one. Separating the two
    files is the prerequisite for fixing that, not the fix.

Tests
    ruff clean on both files; both parse; the class identity above verified at
    runtime. Full-suite verification of the relocation was still running when
    this was committed -- it is the only real proof for a 2,942-line move, and
    any failure it reports will be fixed on this branch.
